% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment.R
\name{create_pipeline_output}
\alias{create_pipeline_output}
\title{Put output of data pipeline step onto s3}
\usage{
create_pipeline_output(output_object, model_name, component, version, type)
}
\arguments{
\item{output_object}{R object representing output of pipeline step.}

\item{model_name}{Name of the model/pipeline.}

\item{component}{Name of pipeline step.}

\item{version}{Version of input.}

\item{type}{File type of output. (i.e. "rds", "txt", "csv")}
}
\value{
string with s3 file path
}
\description{
Properly version and timestamp output of a datapipeline step with associated metadata.  Output needs to exist on local machine.
}
\examples{
create_pipeline_output(training_data_final,"activation_curves", "train_dataprep", 2, "rds")
}
