% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devops.R
\name{create_redshift_con}
\alias{create_redshift_con}
\title{Create Redshift connection object
Read credentials from redshift file, and return valid connection object.}
\usage{
create_redshift_con()
}
\value{
database connection object
}
\description{
Create Redshift connection object
Read credentials from redshift file, and return valid connection object.
}
\examples{
redshift_conn <- create_redshift_con()
}
