% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment.R
\name{feed}
\alias{feed}
\title{Get data dependencies in pipeline from data-deploy script.}
\usage{
feed(model_name, file_name)
}
\arguments{
\item{model_name}{Name of the model/pipeline.}

\item{file_name}{Optional manual override to read in input dataset.}
}
\description{
Parse command line arguments from data-deploy script and return back an input dataset(s) and a fullpath string.
}
\examples{
feed("ops_forecasts_returned_units", "train_dataprep.rds")
}
