% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_functions.R
\name{generate_create_table_sql}
\alias{generate_create_table_sql}
\title{Generate create table DDL from a dataframe}
\usage{
generate_create_table_sql(table_name, df)
}
\arguments{
\item{table_name}{name of table to create on Redshift}

\item{df}{R dataframe object}
}
\value{
string with create table syntax, can be plugged into dbGetQuery
}
\description{
Generates SQL to create a new table from R dataframe- useful in conjuction with copying from s3 to Redshift
}
\examples{
redshift_conn <- create_redshift_con()
sql_create_statement <- generate_create_table_sql("data.activations",activation_data)
dbGetQuery(redshift_conn, sql_create_statement)
}
