% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devops.R
\name{get_aws_credentials}
\alias{get_aws_credentials}
\title{Retrieve AWS credentials}
\usage{
get_aws_credentials()
}
\value{
credentials list containing access key and secret key
}
\description{
Read AWS credentials from local ~/.aws/ directory and store as list. Requires a properly formatted credentials file.
}
\examples{
creds <- get_aws_credentials()
get_bucket(
  bucket = 'everlane-data-deploy',
  key = aws_creds[["aws_access_key_id"]],
  secret = aws_creds[["aws_secret_access_key"]],
  prefix = paste(model_name, component, version, "",sep = "/"),
  max = Inf
)
}
