% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devops.R
\name{get_credentials}
\alias{get_credentials}
\title{Read generic credentials list from file
construct function that grabs credentials from any api -- api creds must always have api name in front of credentials}
\usage{
get_credentials(api)
}
\arguments{
\item{api}{string representing api name- must match prefix of relevant elements in env file}
}
\description{
Read generic credentials list from file
construct function that grabs credentials from any api -- api creds must always have api name in front of credentials
}
\examples{
sailthru_creds <- get_credentials('sailthru') # credentials in env file need to be formatted like sailthru_access_key
}
