% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devops.R
\name{get_redshift_credentials}
\alias{get_redshift_credentials}
\title{Retrieve Redshift credentials}
\usage{
get_redshift_credentials()
}
\value{
credentials list containing host, dbname, user, password, and port
}
\description{
Read Redshift credentials from local ~/.aws/ directory and store as list. Requires a properly formatted redshift file.
}
\examples{
drv <- dbDriver("PostgreSQL")
connection_details <- get_redshift_credentials()
dbConnect(drv,
          host = connection_details['host'],
          dbname = connection_details['dbname'],
          user = connection_details['user'],
          password = connection_details['password'],
          port = connection_details['port'])
}
