% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{tidy_hexbin}
\alias{tidy_hexbin}
\title{Produce x/y coordinates of each hex in a hexbin plot}
\usage{
tidy_hexbin(df, x_column_name, y_column_name)
}
\arguments{
\item{df}{R object containing a tidy dataset}

\item{x_column_name}{Name of column that will represent X axis of graph}

\item{y_column_name}{Name of column that will represent Y axis of graph}
}
\value{
R dataframe with two extra columns representing coordinate of centers for each hex in a hexbin plot
}
\description{
Given a tidy dataset, return a tidy dataset containing x/y coordinates for hexes
}
\examples{
create_pipeline_output(training_data_final,"activation_curves", "train_dataprep", 2, "rds")
}
