% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_functions.R
\name{upload_to_s3}
\alias{upload_to_s3}
\title{Put output of data pipeline step onto s3}
\usage{
upload_to_s3(file_path, object_name, bucket_name = "everlane-data")
}
\arguments{
\item{file_path}{Full path to local filename}

\item{object_name}{Full path to s3 location to upload to}

\item{bucket_name}{Name of s3 bucket}
}
\description{
Easily upload files to S3. Objects to call within function are working file path and object name.
}
\examples{
upload_to_s3("local_file.txt","s3_file_path.txt", bucket_name = "everlane-data")
}
